requirements.txt
numpy 
tensorflow==1.14.0
pillow
opencv-python